V2: Private Todo Lists

+++ index.js
@@ -6,7 +6,7 @@ const {
   getIdeas,
   deleteIdea,
 } = require("./commands/agenda");
-
+const { addTodo, getTodo, deleteTodo } = require("./commands/todo");
 const Discord = require("discord.js");
 const client = new Discord.Client({
   partials: ["MESSAGE", "CHANNEL", "REACTION"],
@@ -25,17 +25,58 @@ client.on("message", async (msg) => {
   }
 
   if (message.startsWith("!addidea")) {
-    const idea = message.substr(message.indexOf(" ") + 1);
-    addIdea(msg, idea);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      const idea = message.substr(message.indexOf(" ") + 1);
+      addIdea(msg, idea);
+    } else {
+      msg.reply("Access denied to command");
+    }
+  }
+
+  if (message.startsWith("!addtodo")) {
+    const messageDetails = message.substr(message.indexOf(" ") + 1);
+    const team = messageDetails.split(" ")[0];
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      addTodo(msg, messageDetails);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!ideas")) {
-    getIdeas(msg, dsc);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      getIdeas(msg, dsc);
+    } else {
+      msg.reply("Access denied to command");
+    }
+  }
+
+  if (message.startsWith("!todo")) {
+    const team = message.substr(message.indexOf(" ") + 1);
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      getTodo(msg, team, dsc);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!removeidea")) {
-    const idea = message.substr(message.indexOf(" ") + 1);
-    deleteIdea(msg, idea);
+    if (msg.member.roles.cache.find((r) => r.name === "Officers")) {
+      const idea = message.substr(message.indexOf(" ") + 1);
+      deleteIdea(msg, idea);
+    } else {
+      msg.reply("Access denied to command");
+    }
+  }
+
+  if (message.startsWith("!removetodo")) {
+    const messageDetails = message.substr(message.indexOf(" ") + 1);
+    const team = messageDetails.split(" ")[0];
+    if (msg.member.roles.cache.some((r) => r.name.toLowerCase() === team)) {
+      deleteTodo(msg, messageDetails);
+    } else {
+      msg.reply("Access denied to command");
+    }
   }
 
   if (message.startsWith("!rules")) {
